Bug fixes and use doubles in cost model to avoid

+++ src/main/scala/nodes/learning/BlockLinearMapper.scala
@@ -267,9 +267,9 @@ class BlockLeastSquaresEstimator[T <: Vector[Double]](blockSize: Int, numIter: I
   override def cost(dataProfile: DataProfile, clusterProfile: ClusterProfile): Double = {
     (dataProfile, clusterProfile) match {
       case (DataProfile(n, d, k, sparsity), ClusterProfile(numMachines, cpuWeight, memWeight, networkWeight)) =>
-        val flops = n * d * (blockSize + k) / numMachines.toDouble
-        val bytesScanned = n * d / numMachines.toDouble + (d * k)
-        val network = 2 * (d * (blockSize + k)) * math.log(numMachines)
+        val flops = n.toDouble * d * (blockSize + k) / numMachines
+        val bytesScanned = n.toDouble * d / numMachines + (d.toDouble * k)
+        val network = 2.0 * (d.toDouble * (blockSize + k)) * math.log(numMachines)
 
         numIter * (math.max(cpuWeight * flops, memWeight * bytesScanned) + networkWeight * network)
     }
